[Optimized] - Removed unused checks for unsigned data type and possible logic error for char type (gcc treats char as unsigned).
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@6814 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/common/socket.c
@@ -1210,7 +1210,7 @@ int RFIFOSKIP(int fd,int len)
 
 	s = session[fd];
 
-	if (s->rdata_size-s->rdata_pos-len<0) {
+	if ((signed int)(s->rdata_size-s->rdata_pos-len)<0) {
 		//fprintf(stderr,"too many skip\n");
 		//exit(1);
 		//better than a COMPLETE program abort // TEST! :)
